+++ jdaviz/configs/specviz/tests/test_helper.py
@@ -405,6 +405,27 @@ def test_load_spectrum_list_directory_concat(tmpdir, specviz_helper):
     assert len(specviz_helper.app.data_collection) == 41
 
 
+def test_load_2d_flux(specviz_helper):
+    # Test loading a spectrum with a 2D flux, which should be split into separate
+    # 1D Spectrum1D objects to load in Specviz.
+    spec = Spectrum1D(spectral_axis=np.linspace(4000, 6000, 10)*u.Angstrom,
+                      flux=np.ones((4, 10))*u.Unit("1e-17 erg / (Angstrom cm2 s)"))
+    specviz_helper.load_data(spec, data_label="test")
+
+    assert len(specviz_helper.app.data_collection) == 4
+    assert specviz_helper.app.data_collection[0].label == "test [0]"
+
+    spec2 = Spectrum1D(spectral_axis=np.linspace(4000, 6000, 10)*u.Angstrom,
+                       flux=np.ones((2, 10))*u.Unit("1e-17 erg / (Angstrom cm2 s)"))
+
+    # Make sure 2D spectra in a SpectrumList also get split properly.
+    spec_list = SpectrumList([spec, spec2])
+    specviz_helper.load_data(spec_list, data_label="second test")
+
+    assert len(specviz_helper.app.data_collection) == 10
+    assert specviz_helper.app.data_collection[-1].label == "second test [5]"
+
+
 def test_plot_uncertainties(specviz_helper, spectrum1d):
     specviz_helper.load_data(spectrum1d)